refactor: extract hardcoded values to constants
This commit addresses the "Constants & Configuration" phase of the AI Slop Reduction roadmap.

- Extracted magic numbers and hardcoded configuration values from various components and hooks, including `useBluetoothHRM`, `socketManager`, and `websocketUtils`.
- Centralized these values into new, well-scoped constant files (`lib/constants/index.ts` and `hooks/useBluetoothHRM.constants.ts`).
- This improves code quality, reduces duplication, and makes the system easier for both humans and AI to understand and maintain.
- Resolved a build failure caused by conflicting module resolution between the `tsc` server build and the Next.js client build by moving shared constants into the `lib/` directory.
- Fixed a failing unit test (`hr-zones.test.ts`) that was using outdated hardcoded values, ensuring it now correctly uses the centralized constants.
