+++ hooks/useBluetoothHRM.ts
@@ -12,7 +12,7 @@ import {
 } from '../types/websocket'
 import { BluetoothConnectionStatus } from '../types/bluetooth'
 import isEqual from 'lodash.isequal'
-import { calculateMaxHr } from '../utils/constants'
+import { calculateMaxHr } from '@/lib/constants'
 import logger from '@/utils/logger'
 import { useWebSocket } from '@/context/WebSocketContext'
 import { cancellablePromise } from '@/utils/promise'
